+++ src/app/dashboard/compounds/page.tsx
@@ -1,11 +1,12 @@
-import React from 'react'
+import React from 'react';
+import CompoundBatchesList from '@/components/compound-batches';
 
 const page = () => {
   return (
-      <div>
-          This is compounds page
+    <div className="p-6">
+      <CompoundBatchesList />
     </div>
-  )
-}
+  );
+};
 
-export default page
+export default page;